[#44] Refactor: 일정 추가 모달 달려 컴포넌트 분리, 달력 배열 함수 분리
