+++ components/calendar/calendar-modal-calendar.module.scss
@@ -1,0 +1,39 @@
+.container {
+  display: grid;
+  grid: auto-flow / repeat(7, 1fr);
+  width: 100%;
+}
+
+.calendarCell {
+  text-align: center;
+  font-size: 16px;
+  font-weight: 500;
+  color: var(--text-color-gray);
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+  gap: 8px;
+  padding: 8px;
+  cursor: pointer;
+
+  &.red {
+    color: var(--calendar-red);
+  }
+
+  .date {
+    font-size: 13px;
+    font-weight: 600;
+    width: 20px;
+    height: 20px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+
+  .selected {
+    background-color: var(--main-color);
+    border-radius: 9999px;
+    color: var(--white);
+  }
+}
